Added new pages
-issues
-projects
-tasks
