+++ index.html
@@ -27,7 +27,7 @@
             </a>
           </li>
           <li class="nav-item">
-            <a href="#tasks" class="nav-link">
+            <a href="tasks.html" class="nav-link">
               <img src="src/assets/icons/tasks.svg" alt="" loading="lazy"/>
               <p>Tasks</p>
             </a>
@@ -41,7 +41,7 @@
         <nav class="navigation">
           <ul class="nav-list">
             <li class="nav-item">
-              <a href="#all-projects" class="nav-link">
+              <a href="projects.html" class="nav-link">
                 <img src="src/assets/icons/all-projects.svg" alt="" loading="lazy" />
                 <p>All Projects</p>
               </a>
@@ -81,7 +81,7 @@
             </a>
           </li>
           <li class="nav-item">
-            <a href="#issues" class="nav-link">
+            <a href="issues.html" class="nav-link">
               <img src="src/assets/icons/issues.svg" alt="" loading="lazy"/>
               <p>Issues</p>
             </a>
